+++ src/types.ts
@@ -284,6 +284,7 @@ export interface AssetRiskState {
   maintenanceMarginIncludingOrders: number;
   unrealizedPnl: number;
   unpaidFunding: number;
+  potentialOrderLoss: number;
 }
 
 export interface CrossMarginAccountState {
@@ -300,6 +301,7 @@ export interface CrossMarginAccountState {
   maintenanceMarginIncludingOrdersTotal: number;
   unrealizedPnlTotal: number;
   unpaidFundingTotal: number;
+  potentialOrderLossTotal: number;
 }
 
 export interface CancelArgs {